Re-enable test of `composer` `Dependabot::PrivateSourceTimedOut`
This was commented out because it had to wait til `composer` timed out,
which defaults to 300 seconds (5 mins).

However, `composer` supports setting a custom timeout using the env var
`COMPOSER_PROCESS_TIMEOUT`.

So set a much lower timeout and re-enable the test. When I first
re-enabled the test, I discovered a code bug. Furthermore, once that was
fixed, I discovered a bug in the test itself, so we should keep this
enabled if we have any hope of deriving value from it.