Update ignored since `3.0-dev` works on php 8
This is for the `monolog/monolog` dep, which switched in v3.0 to
requiring php 8.

So under php `7.4`, the test would never pull `3.0`, so ignored the
latest version.

Confusingly, the test used to return `2.8.0`, when it should have
returned `1.21.0` because `2.8.0` is within the version range that's
being ignored.

In fact, according to https://github.com/dependabot/dependabot-core/pull/6334/commits/34216d42d099c60d9d999664e4dcb956721e1c89,
it used to return `1.21.0`... I'm not sure why when I updated the
packagist response in that commit it started returning `2.8.0` instead...
But not worth digging into since the test behavior is now what we
expect... `1.21.0` is the newest version that isn't ignored.

squashme