Remove `EndToEndJob`
This is an old artifact from before `FileFetcherJob` and `UpdateFilesJob`
existed...

We've long wanted to get rid of these, there are at least two PR's I
found in our internal repo from before we open sourced this code.

But the blocker was that we didn't have integration tests that tested
the combination of `FileFetcherJob` + `UpdateFilesJob`, so we wanted
to refactor the integration tests to test the handoff between the two.

To be clear: We _do_ have unit tests that test each of those individual
components, so we can still refactor internally knowing that we're
protected by individual unit tests for a fast dev/test cycle. It's just
the combination of the two that's not tested.

However, we now have the full `smoke` test suite which tests everything
end to end. So rather than refactoring the integration tests, it's now
safe to simply remove them.